Add 4 Variant

+++ app/src/main/java/com/example/doghotel/model/Dog.kt
@@ -12,7 +12,7 @@ data class Dog(
     @ColumnInfo (name = "nickname") var nickname: String = "",
     @ColumnInfo (name = "gender") var gender: String = "",
     @ColumnInfo (name = "age") var age: Int = 1,
-    @ColumnInfo (name = "days") var days: Long = 1L ,
+    @ColumnInfo (name = "days") var days: Int = 1 ,
     @ColumnInfo (name = "cage") var cage: Int = 0,
     @ColumnInfo (name = "photo") var photo: String = ""
